+++ packages/expo-router/src/rsc/router/client.ts
@@ -365,6 +365,9 @@ export function useRouter_UNSTABLE(): ClassicExpoRouterType &
     setParams() {
       throw new Error('router.setParams() is not supported in React Server Components yet');
     },
+    setTransitionMode() {
+      throw new Error('router.setTransitionMode() is not supported in React Server Components yet');
+    },
 
     // TODO: The behavior here is not the same as before.
     navigate: push,